added get route for returning a villain by slug
