+++ controller/villains.js
@@ -6,6 +6,15 @@ const getAllVillains = async (request, response) => {
   return response.send(villains)
 }
 
+const getVillain = async (request, response) => {
+  const { slug } = request.params
+
+  const villain = await models.villains.findOne({ where: { slug } })
+
+  return villain ? response.send(villain) : response.sendStatus(404)
+}
+
 module.exports = {
-  getAllVillains
+  getAllVillains,
+  getVillain
 }
