+++ models/villains.js
@@ -1,9 +1,11 @@
-const Sequelize = require('sequelize')
-const villainsModel = require('./villains')
+const villains = (connection, Sequelize) => {
+  return connection.define('villains',
+    {
+      id: { type: Sequelize.INTEGER, autoIncrement: true, primaryKey: true },
+      name: { type: Sequelize.STRING },
+      movie: { type: Sequelize.STRING },
+      slug: { type: Sequelize.STRING },
+    }, { paranoid: true })
+}
 
-const connection = new Sequelize('disney', 'disneyUser', 'password123',
-  { host: 'localhost', dialect: 'mysql' })
-
-const villains = villainsModel(connection, Sequelize)
-
-module.exports = { villains }
+module.exports = villains
